pyproject.toml
[project]
name = "dspy-pdf"
version = "0.1.0"
description = "Extract pinout information from electronics datasheets using DSPy and RAG"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["dspy", "pdf", "pinout", "datasheet", "rag", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
dependencies = [
    "chromadb>=1.0.15",
    "dspy-ai>=2.6.27",
    "llama-index>=0.12.50",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-vector-stores-chroma>=0.4.2",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
]

[project.scripts]
dspy-pdf = "main:main"
